Move misc to common
