+++ lahja/__init__.py
@@ -1,9 +1,9 @@
 from .asyncio import AsyncioEndpoint  # noqa: F401
-from .exceptions import ConnectionAttemptRejected, UnexpectedResponse  # noqa: F401
-from .misc import (  # noqa: F401
+from .common import (  # noqa: F401
     BaseEvent,
     BaseRequestResponseEvent,
     BroadcastConfig,
     ConnectionConfig,
     Subscription,
 )
+from .exceptions import ConnectionAttemptRejected, UnexpectedResponse  # noqa: F401
