+++ lahja/common.py
